+++ changelog.md
@@ -2,11 +2,13 @@
 
 ## 0.11.3
 
+- add `events` to task context, so tasks can communicate with each other using a normal `EventEmitter`
+  ([#143](https://github.com/feltcoop/gro/pull/143))
+  - events aren't great for everything;
+    this PR also documents a value mapping pattern convention for tasks in `src/task/README.md`
 - `gro build` now correctly builds only `BuildConfig`s that have `dist: true`,
   allowing users to customize the `dist/` output in each `gro build` via `src/gro.config.ts`
   ([#144](https://github.com/feltcoop/gro/pull/144))
-- add utils for callback hooks at `src/utils/listener.ts`
-  ([#143](https://github.com/feltcoop/gro/pull/143))
 
 ## 0.11.2